Move roadmap to the landing page layout, phase out updates timeline
