+++ docs/app/(site)/roadmap/page.tsx
@@ -1,0 +1,10 @@
+import PageClient from './page-client'
+
+export const metadata = {
+  title: 'Roadmap',
+  description: "Discover where KeystoneJS is headed, and why we're going there.",
+}
+
+export default function ForOrganisations () {
+  return <PageClient />
+}
